Use a more efficient TableReader::get() implementation.
The old implementation was simply based on iterators and seeking; the new one
is "native" in that it implements the details itself, using the index blocks
for a rough seek, the filter blocks for skipping irrelevant blocks, and linear
search within a block for the key itself.

This is a code rebase from leveldb.