Add `current_key()` function to SSIterator.
In case we are not interested in the actual value, but just want to check that an entry exists, we only need access to the key and can save the time to copy the (possible large) value. In addition, returning the key as a reference to an already existing field allows even more optimized acccess.