Allow to set the block cache size manually.
While the other configuration options are marked as public an can be tuned manually,
the Cache type is not public and therefore can't be constructed from outside the crate.
This adds a "with_cache_capacity" function which allows to change the options
to use a block cache with a custom size.